ENH: Addition of the tensor correction flag
When computing and resampling tensors, some tensors may end up having negative eigen-values. To avoid these bad tensors, we apply a post correction (absolute value of negative eigen-values, negative values set to zero, or tensor replaced by nearest tensor [1]). This post-correction option was already available in ResampleDTIlogEuclidean, but was not available in dtiprocess.
The option has been added to dtiprocess and is now exposed in DTI-Reg.

[1] Nicholas J. Higham. Computing a nearest symmetric positive semidefinite matrix. Linear Algebra and its Applications, 103:103–118, May 1988.