ENH: Possibility to compute the MD as the scalar measurement instead of the FA
When registering diffusion tensor images, registering the Fractional Anisotropy (FA) typically works well. However, this is not the case when registering diffusion images of small animals such as rats and mice. In this case, using the Mean Diffusivity (MD) gives better results. An option has been added to DTI-Reg to be able to select either the FA or the MD as the scalar measurement that is used to drive the registration. The default is still the FA.