ENH: gaussianSmoothingSigmas is not necessary anymore
In earlier versions of ANTS, if --gaussian-smoothing-sigmas was not specified when calling ANTs, a warning message was printed out. To avoid printing out this message, DTI-Reg was generating default values (4x2x1 if 3 levels of registration) to smooth the images. However, this was creating issues since these values were not adapted to a spacing different from 1x1x1mm^3 (especially the case with monkeys or rodents). In newer versions of ANTs, this warning message is not printed, and default values adapted to the scale of the image are computed by ANTs. Therefore we removed the code that was generating the default values in DTI-Reg.